Update v6
On the home page, make the title of a thread clickable. When users click the title, they will visit the thread page.

I'll help you modify the code to make thread titles clickable and add a dedicated thread page view. This will involve:
Modifying the render function to make the title clickable
Adding a new route and page for viewing individual threads
First, I'll modify the render function to wrap the title in a clickable link:
Now I'll add a new route and page for viewing individual threads:
I've made the following changes:
Modified the render function to make the thread title clickable by wrapping it in an A tag with hx_get pointing to /thread/{thread.id}. I've styled it to look like a regular heading while still being clickable.
Added a new route /thread/{tid} that displays a dedicated page for viewing individual threads. This page includes:
A back button to return to the home page
The thread title as a large heading
The thread content
The posting timestamp
Edit and Delete buttons with the same functionality as on the home page
The changes use HTMX for smooth page transitions without full page reloads. When users click a thread title, they'll be taken to a dedicated page for that thread, and they can easily navigate back to the home page using the back button.